+++ .pnp.cjs
@@ -759,6 +759,10 @@ const RAW_RUNTIME_STATE =
       "reference": "workspace:packages/sources/synthetix-debt-pool"\
     },\
     {\
+      "name": "@chainlink/synthetix-feeds-adapter",\
+      "reference": "workspace:packages/sources/synthetix-feeds"\
+    },\
+    {\
       "name": "@chainlink/taapi-adapter",\
       "reference": "workspace:packages/sources/taapi"\
     },\
@@ -1064,6 +1068,7 @@ const RAW_RUNTIME_STATE =
     ["@chainlink/swell-address-list-adapter", ["workspace:packages/sources/swell-address-list"]],\
     ["@chainlink/synth-index-adapter", ["workspace:packages/composites/synth-index"]],\
     ["@chainlink/synthetix-debt-pool-adapter", ["workspace:packages/sources/synthetix-debt-pool"]],\
+    ["@chainlink/synthetix-feeds-adapter", ["workspace:packages/sources/synthetix-feeds"]],\
     ["@chainlink/taapi-adapter", ["workspace:packages/sources/taapi"]],\
     ["@chainlink/terra-view-function-adapter", ["workspace:packages/sources/terra-view-function"]],\
     ["@chainlink/the-graph-adapter", ["workspace:packages/composites/the-graph"]],\
@@ -9231,6 +9236,23 @@ const RAW_RUNTIME_STATE =
           ["@types/supertest", "npm:2.0.16"],\
           ["ethers", "npm:5.7.2"],\
           ["supertest", "npm:6.2.4"],\
+          ["tslib", "npm:2.4.1"],\
+          ["typescript", "patch:typescript@npm%3A5.0.4#optional!builtin<compat/typescript>::version=5.0.4&hash=b5f058"]\
+        ],\
+        "linkType": "SOFT"\
+      }]\
+    ]],\
+    ["@chainlink/synthetix-feeds-adapter", [\
+      ["workspace:packages/sources/synthetix-feeds", {\
+        "packageLocation": "./packages/sources/synthetix-feeds/",\
+        "packageDependencies": [\
+          ["@chainlink/synthetix-feeds-adapter", "workspace:packages/sources/synthetix-feeds"],\
+          ["@chainlink/external-adapter-framework", "npm:1.3.1"],\
+          ["@types/jest", "npm:27.5.2"],\
+          ["@types/node", "npm:16.11.68"],\
+          ["decimal.js", "npm:10.4.3"],\
+          ["ethers", "npm:5.7.2"],\
+          ["nock", "npm:13.5.4"],\
           ["tslib", "npm:2.4.1"],\
           ["typescript", "patch:typescript@npm%3A5.0.4#optional!builtin<compat/typescript>::version=5.0.4&hash=b5f058"]\
         ],\