+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.78
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.12.5
+
 ## 3.0.77
 
 ### Patch Changes