+++ packages/composites/synth-index/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/synth-index-adapter
 
+## 1.2.97
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.12.5
+
 ## 1.2.96
 
 ### Patch Changes